nvim: remap <Leader>{y,p} to clipboard selection.
Mouse selection is accessible with capitalized Y and P: <Leader>{Y,P}.
The intention is to make things copied with a keyboard easier accessible
with a keyboard. Mouse selection you can insert with mouse. If you
copied it with mouse, than mouse is nearby and you can insert with mouse
too.